added functionality to delete files from db and server

+++ backend/src/image/service/image.service.ts
@@ -4,7 +4,7 @@ import { Repository } from "typeorm";
 import { InjectRepository } from "@nestjs/typeorm";
 import { ImageEntity } from "../model/domain/image.entity";
 import { ImageDto } from "../model/dto/image.dto";
-import { mapImageToImageDto } from "../util/util.functions";
+import { deleteFileFromDisk, mapImageToImageDto } from "../util/util.functions";
 import { ImageDownloadException } from "../../exceptions/type/image.download.exception";
 import * as process from "process";
 
@@ -21,7 +21,7 @@ export class ImageService {
 
   async storeFiles(files: Array<Express.Multer.File>): Promise<ImageDto[]> {
     const images: ImageEntity[] = [];
-    this.logger.debug(`store ${files.length} files`);
+    this.logger.debug(`Store ${files.length} files`);
     try {
       for (const file of files) {
         const image = new ImageEntity(file.filename, file.mimetype);
@@ -32,12 +32,12 @@ export class ImageService {
       this.logger.error(error.message);
       throw new ImageDownloadException(error.message);
     }
-    this.logger.debug('successfully stored files');
+    this.logger.debug('Successfully stored files');
     return images.map(image => mapImageToImageDto(image));
   }
 
   async getFile(id: number): Promise<ImageEntity> {
-    this.logger.debug(`get file with id ${id}`);
+    this.logger.debug(`Get file with id ${id}`);
     const file: ImageEntity = await this.imageRepository.findOneBy({ id: id });
     if (!file) {
       throw new ImageDownloadException(`The file with the id ${id} does not exist.`);
@@ -46,14 +46,16 @@ export class ImageService {
   }
 
   async deleteFile(id: number): Promise<boolean> {
-    this.logger.debug(`delete file with id ${id}`);
-    try {
-      await this.imageRepository.delete({ id: id });
-      return Promise.resolve(true);
-    } catch (error) {
-      this.logger.error(error.message);
-      throw new ImageDownloadException(error.message);
+    this.logger.debug(`Attempt to delete file with ID: ${id}`);
+    const file: ImageEntity = await this.imageRepository.findOneBy({ id: id });
+    if (!file) {
+      this.logger.error(`The file with the id ${id} does not exist.`);
+      throw new ImageDownloadException(`The file with the id ${id} does not exist.`);
     }
+    await this.imageRepository.delete({ id: id });
+    const path = `${process.env.MULTER_STORAGE_PATH}\\${file.filename}`;
+    deleteFileFromDisk(path);
+    return Promise.resolve(true);
   }
 
   async getFileBuffer(file: ImageEntity) {
